% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aritmetica.R
\name{aritmetica}
\alias{aritmetica}
\title{Nome pacote jo arquivo .R}
\usage{
aritmetica(a, b, op)
}
\arguments{
\item{a}{primeira variavel}

\item{b}{segunda variavel}

\item{op}{operacao matematica}
}
\value{

}
\description{
Nome pacote jo arquivo .R
}
